Refactor AppDetails component into separate file
